Refactor community menu options
